feat: move types to shared folder
