+++ src/modules/shared/types/newsletter.ts
@@ -1,0 +1,14 @@
+export interface NewsletterFormState {
+	firstName: string;
+}
+
+export interface GetNewsletterPreferencesResponse {
+	newsletter: boolean;
+}
+
+export interface SetNewsletterPreferencesBody {
+	firstName?: string;
+	lastName?: string;
+	mail?: string;
+	preferences: GetNewsletterPreferencesResponse;
+}
